♻️ REFACTOR: Move implementation specific attribute code
`BackendEntityAttributesMixin` leaked implementation specific variables and functions.
This has now been moved to the `DjangoNode`/`SqlaNode` classes.
