+++ aiida/orm/implementation/entities.py
@@ -16,9 +16,7 @@ from aiida.orm.implementation.utils import clean_value, validate_attribute_extra
 if TYPE_CHECKING:
     from aiida.orm.implementation import Backend
 
-__all__ = (
-    'BackendEntity', 'BackendCollection', 'EntityType', 'BackendEntityAttributesMixin', 'BackendEntityExtrasMixin'
-)
+__all__ = ('BackendEntity', 'BackendCollection', 'EntityType', 'BackendEntityExtrasMixin')
 
 EntityType = TypeVar('EntityType', bound='BackendEntity')  # pylint: disable=invalid-name
 
@@ -111,171 +109,6 @@ class BackendCollection(Generic[EntityType]):
         :return: the newly created entry of type ENTITY_CLASS
         """
         return self.ENTITY_CLASS(backend=self._backend, **kwargs)
-
-
-class BackendEntityAttributesMixin(abc.ABC):
-    """Mixin class that adds all methods for the attributes column to a backend entity"""
-
-    _dbmodel: Any
-    dbmodel: Any
-    is_stored: Any
-
-    def _flush_if_stored(self, fields: Set[str]) -> None:
-        if self._dbmodel.is_saved():
-            self._dbmodel._flush(fields)  # pylint: disable=protected-access
-
-    @property
-    def attributes(self):
-        """Return the complete attributes dictionary.
-
-        .. warning:: While the entity is unstored, this will return references of the attributes on the database model,
-            meaning that changes on the returned values (if they are mutable themselves, e.g. a list or dictionary) will
-            automatically be reflected on the database model as well. As soon as the entity is stored, the returned
-            attributes will be a deep copy and mutations of the database attributes will have to go through the
-            appropriate set methods. Therefore, once stored, retrieving a deep copy can be a heavy operation. If you
-            only need the keys or some values, use the iterators `attributes_keys` and `attributes_items`, or the
-            getters `get_attribute` and `get_attribute_many` instead.
-
-        :return: the attributes as a dictionary
-        """
-        return self._dbmodel.attributes
-
-    def get_attribute(self, key):
-        """Return the value of an attribute.
-
-        .. warning:: While the entity is unstored, this will return a reference of the attribute on the database model,
-            meaning that changes on the returned value (if they are mutable themselves, e.g. a list or dictionary) will
-            automatically be reflected on the database model as well. As soon as the entity is stored, the returned
-            attribute will be a deep copy and mutations of the database attributes will have to go through the
-            appropriate set methods.
-
-        :param key: name of the attribute
-        :return: the value of the attribute
-        :raises AttributeError: if the attribute does not exist
-        """
-        try:
-            return self._dbmodel.attributes[key]
-        except KeyError as exception:
-            raise AttributeError(f'attribute `{exception}` does not exist') from exception
-
-    def get_attribute_many(self, keys):
-        """Return the values of multiple attributes.
-
-        .. warning:: While the entity is unstored, this will return references of the attributes on the database model,
-            meaning that changes on the returned values (if they are mutable themselves, e.g. a list or dictionary) will
-            automatically be reflected on the database model as well. As soon as the entity is stored, the returned
-            attributes will be a deep copy and mutations of the database attributes will have to go through the
-            appropriate set methods. Therefore, once stored, retrieving a deep copy can be a heavy operation. If you
-            only need the keys or some values, use the iterators `attributes_keys` and `attributes_items`, or the
-            getters `get_attribute` and `get_attribute_many` instead.
-
-        :param keys: a list of attribute names
-        :return: a list of attribute values
-        :raises AttributeError: if at least one attribute does not exist
-        """
-        try:
-            return [self.get_attribute(key) for key in keys]
-        except KeyError as exception:
-            raise AttributeError(f'attribute `{exception}` does not exist') from exception
-
-    def set_attribute(self, key, value):
-        """Set an attribute to the given value.
-
-        :param key: name of the attribute
-        :param value: value of the attribute
-        """
-        validate_attribute_extra_key(key)
-
-        if self.is_stored:
-            value = clean_value(value)
-
-        self._dbmodel.attributes[key] = value
-        self._flush_if_stored({'attributes'})
-
-    def set_attribute_many(self, attributes):
-        """Set multiple attributes.
-
-        .. note:: This will override any existing attributes that are present in the new dictionary.
-
-        :param attributes: a dictionary with the attributes to set
-        """
-        for key in attributes:
-            validate_attribute_extra_key(key)
-
-        if self.is_stored:
-            attributes = {key: clean_value(value) for key, value in attributes.items()}
-
-        for key, value in attributes.items():
-            # We need to use `self.dbmodel` without the underscore, because otherwise the second iteration will refetch
-            # what is in the database and we lose the initial changes.
-            self.dbmodel.attributes[key] = value
-        self._flush_if_stored({'attributes'})
-
-    def reset_attributes(self, attributes):
-        """Reset the attributes.
-
-        .. note:: This will completely clear any existing attributes and replace them with the new dictionary.
-
-        :param attributes: a dictionary with the attributes to set
-        """
-        for key in attributes:
-            validate_attribute_extra_key(key)
-
-        if self.is_stored:
-            attributes = clean_value(attributes)
-
-        self.dbmodel.attributes = attributes
-        self._flush_if_stored({'attributes'})
-
-    def delete_attribute(self, key):
-        """Delete an attribute.
-
-        :param key: name of the attribute
-        :raises AttributeError: if the attribute does not exist
-        """
-        try:
-            self._dbmodel.attributes.pop(key)
-        except KeyError as exception:
-            raise AttributeError(f'attribute `{exception}` does not exist') from exception
-        else:
-            self._flush_if_stored({'attributes'})
-
-    def delete_attribute_many(self, keys):
-        """Delete multiple attributes.
-
-        :param keys: names of the attributes to delete
-        :raises AttributeError: if at least one of the attribute does not exist
-        """
-        non_existing_keys = [key for key in keys if key not in self._dbmodel.attributes]
-
-        if non_existing_keys:
-            raise AttributeError(f"attributes `{', '.join(non_existing_keys)}` do not exist")
-
-        for key in keys:
-            self.dbmodel.attributes.pop(key)
-
-        self._flush_if_stored({'attributes'})
-
-    def clear_attributes(self):
-        """Delete all attributes."""
-        self._dbmodel.attributes = {}
-        self._flush_if_stored({'attributes'})
-
-    def attributes_items(self):
-        """Return an iterator over the attributes.
-
-        :return: an iterator with attribute key value pairs
-        """
-        for key, value in self._dbmodel.attributes.items():
-            yield key, value
-
-    def attributes_keys(self):
-        """Return an iterator over the attribute keys.
-
-        :return: an iterator with attribute keys
-        """
-        for key in self._dbmodel.attributes.keys():
-            yield key
 
 
 class BackendEntityExtrasMixin(abc.ABC):
